--- client/store/cart.js
@@ -1,24 +1,0 @@
-import axios from 'axios'
-
-const ADDED_ART = 'ADDED_ART'
-
-const addedArt = art => ({
-  type: ADDED_ART,
-  art
-})
-
-export const addArt = artId => {
-  return async dispatch => {
-    try {
-      const {data} = await axios.put(`/api/order/add/${artId}`)
-      dispatch(addedArt(data))
-    } catch (error) {
-      let myData = JSON.parse(localStorage.getItem('cart')) || {arts: []}
-      if (!myData.arts.map(x => x.id).includes(artId)) {
-        const {data} = await axios.get(`/api/art/${artId}`)
-        myData.arts = myData.arts.concat(data)
-        localStorage.setItem('cart', JSON.stringify(myData))
-      }
-    }
-  }
-}
