+++ client/store/cartReducer.js
@@ -3,6 +3,7 @@ import Axios from 'axios'
 //Action Type
 const ALL_CART = 'ALL_CART'
 const CHECK_OUT = 'CHECK_OUT'
+const ADDED_ART = 'ADDED_ART'
 
 //Action Creator
 const gotCart = cart => ({
@@ -13,15 +14,19 @@ const checkOut = cart => ({
   type: CHECK_OUT,
   cart
 })
+const addedArt = art => ({
+  type: ADDED_ART,
+  art
+})
+
 //Thunk Creator
 export const getCartThunk = () => {
   return async dispatch => {
     try {
       const {data} = await Axios.get('/api/order/cart')
-      if (data) dispatch(gotCart(data))
-      else dispatch(gotCart(window.localStorage.getItem('cart')))
+      dispatch(gotCart(data))
     } catch (error) {
-      console.log(error)
+      dispatch(gotCart(JSON.parse(localStorage.getItem('cart'))))
     }
   }
 }
@@ -35,6 +40,21 @@ export const checkOutThunk = () => {
     }
   }
 }
+export const addArt = artId => {
+  return async dispatch => {
+    try {
+      const {data} = await Axios.put(`/api/order/add/${artId}`)
+      dispatch(addedArt(data))
+    } catch (error) {
+      let myData = JSON.parse(localStorage.getItem('cart')) || {arts: []}
+      if (!myData.arts.map(x => x.id).includes(artId)) {
+        const {data} = await Axios.get(`/api/art/${artId}`)
+        myData.arts = myData.arts.concat(data)
+        localStorage.setItem('cart', JSON.stringify(myData))
+      }
+    }
+  }
+}
 
 //Cart Reducer
 const cartReducer = (state = {}, action) => {
@@ -43,6 +63,8 @@ const cartReducer = (state = {}, action) => {
       return {...state, cart: action.cart}
     case CHECK_OUT:
       return {...state, cart: action.cart}
+    case ADDED_ART:
+      return state
     default:
       return state
   }
