+++ server/api/ProductRoutes.js
@@ -20,7 +20,6 @@ router.get('/', async (req, res, next) => {
 router.get('/:id', async (req, res, next) => {
   try {
     const singleProduct = await Art.findByPk(req.params.id)
-    console.log('SINGLE PRODUCT', singleProduct)
     res.json(singleProduct)
   } catch (error) {
     next(error)
